+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added faster `ex.BoundingBox.transform(...)` implementation.
 - Added faster `ex.BoundingBox.overlap(...)` implementation.
 - Added `ex.Vector.min(...)` and `ex.Vector.max(...)` to find the min/max of each vector component between 2 vectors.
+- Added `ex.TransformComponent.zIndexChange$` observable to watch when z index changes.
 
 ### Fixed